Reduce to metadata tab
