+++ .github/workflows/test_and_deploy.yml
@@ -40,7 +40,7 @@ jobs:
       - name: Cache Test Data
         uses: actions/cache@v3
         with:
-          path: ${{ github.workspace }}/.WAZP/*
+          path: ${{ github.workspace }}/.beekeeper/*
           key: cached-test-data
           enableCrossOsArchive: true
