+++ README.md
@@ -1,35 +1,38 @@
 [![License](https://img.shields.io/badge/License-BSD_3--Clause-orange.svg)](https://opensource.org/licenses/BSD-3-Clause)
-![CI](https://img.shields.io/github/actions/workflow/status/SainsburyWellcomeCentre/WAZP/test_and_deploy.yml?label=CI)
-[![codecov](https://codecov.io/gh/SainsburyWellcomeCentre/WAZP/branch/main/graph/badge.svg?token=DYQB8EHTJB)](https://codecov.io/gh/SainsburyWellcomeCentre/WAZP)
-[![docs](https://img.shields.io/website?down_color=red&down_message=down&label=docs&up_color=brightgreen&up_message=up&url=https%3A%2F%2Fsainsburywellcomecentre.github.io%2FWAZP%2F)](https://sainsburywellcomecentre.github.io/WAZP/)
+![CI](https://img.shields.io/github/actions/workflow/status/SainsburyWellcomeCentre/beekeeper/test_and_deploy.yml?label=CI)
+[![codecov](https://codecov.io/gh/SainsburyWellcomeCentre/beekeeper/branch/main/graph/badge.svg?token=DYQB8EHTJB)](https://codecov.io/gh/SainsburyWellcomeCentre/beekeeper)
+[![docs](https://img.shields.io/website?down_color=red&down_message=down&label=docs&up_color=brightgreen&up_message=up&url=https%3A%2F%2Fsainsburywellcomecentre.github.io%2Fbeekeeper%2F)](https://sainsburywellcomecentre.github.io/beekeeper/)
 [![Ruff](https://img.shields.io/endpoint?url=https://raw.githubusercontent.com/charliermarsh/ruff/main/assets/badge/v0.json)](https://github.com/charliermarsh/ruff)
 [![Code style: black](https://img.shields.io/badge/code%20style-black-000000.svg)](https://github.com/python/black)
 [![pre-commit](https://img.shields.io/badge/pre--commit-enabled-brightgreen?logo=pre-commit&logoColor=white)](https://github.com/pre-commit/pre-commit)
 
-# WAZP 🐝
-**W**asp **A**nimal-tracking **Z**oo project with **P**ose estimation
-(name is subject to refinement)
+# beekeeper 🐝
+Managing video metadata for animal behaviour experiments.
 
-## Overview
+<!-- ## Overview
 
-WAZP is a dashboard built with [Dash-Plotly](https://dash.plotly.com/) for analysing animal tracking data. It can display pose estimation output from [DeepLabCut](https://github.com/DeepLabCut/DeepLabCut).
+beekeeper is a dashboard built with [Dash-Plotly](https://dash.plotly.com/) for analysing animal tracking data. It can display pose estimation output from [DeepLabCut](https://github.com/DeepLabCut/DeepLabCut). -->
 
 The package is currently in early development 🏗️ and is not yet ready for use. Stay tuned ⌛
 
+It is based on an earlier codebase called [WAZP](https://sainsburywellcomecentre.github.io/WAZP/).
+
 ## Installation
 
-We recommend you install WAZP inside a [conda](https://docs.conda.io/en/latest/) environment.
-Once you have `conda` installed, the following will create and activate an environment. You can call your environment whatever you like, we've used `wazp-env`.
+We recommend you install `beekeeper` inside a [conda](https://docs.conda.io/en/latest/) environment.
+Once you have `conda` installed, the following will create and activate an environment. You can call your environment whatever you like, we've used `beekeeper-env`.
 
 ```sh
-conda create -n wazp-env -c conda-forge python=3 pytables
-conda activate wazp-env
+conda create -n beekeeper-env -c conda-forge python=3 pytables
+conda activate beekeeper-env
 ```
 
-Next install the latest version of WAZP from pip:
+Next install the latest version of `beekeeper` from pip:
 
 ```sh
-pip install wazp
+git clone https://github.com/sfmig/beekeeper.git
+cd beekeeper
+pip install .
 ```
 
 ## Launching the dashboard
@@ -37,7 +40,7 @@ pip install wazp
 Once installed, launch the dashboard by running the following command from the root of the repository:
 
 ```sh
-startwazp
+start-beekeeper
 ```
 This will launch a local web server. If the dashboard does not automatically open in your default browser, click the link in the terminal to open it (the link will be of the form `http://localhost:8050/`).
