+++ beekeeper/__init__.py
@@ -1,7 +1,7 @@
 from importlib.metadata import PackageNotFoundError, version
 
 try:
-    __version__ = version("wazp")
+    __version__ = version("beekeeper")
 except PackageNotFoundError:
     # package is not installed
     pass
