+++ beekeeper/app.py
@@ -4,10 +4,8 @@ import dash
 import dash_bootstrap_components as dbc
 from dash import Dash, dcc, html
 
-import wazp.callbacks.dashboard as dashboard
-import wazp.callbacks.home as home
-import wazp.callbacks.metadata as metadata
-import wazp.callbacks.roi as roi
+import beekeeper.callbacks.home as home
+import beekeeper.callbacks.metadata as metadata
 
 #################
 # Initialise app
@@ -16,7 +14,7 @@ app = Dash(
     __name__,
     use_pages=True,
     external_stylesheets=[dbc.themes.BOOTSTRAP],
-    suppress_callback_exceptions=True
+    suppress_callback_exceptions=True,
     # TODO: is there an alternative to prevent error w/ chained callbacks?
 )
 
@@ -37,17 +35,17 @@ SIDEBAR_STYLE = {
 # Sidebar component
 sidebar = html.Div(
     [
-        html.H2("WAZP 🐝", className="display-4"),
+        html.H2("beekeeper 🐝", className="display-4"),
         html.Hr(),
         html.P(
             [
-                "Wasp",
+                "Managing",
                 html.Br(),
-                "Animal-tracking",
+                "video metadata",
                 html.Br(),
-                "Zoo project with",
+                "for animal behaviour",
                 html.Br(),
-                "Pose estimation",
+                "experiments",
             ],
             className="lead",
         ),
@@ -107,11 +105,9 @@ app.layout = html.Div(
 ################
 home.get_callbacks(app)
 metadata.get_callbacks(app)
-roi.get_callbacks(app)
-dashboard.get_callbacks(app)
 
 
-def startwazp():
+def startbeekeeper():
     app.run_server(debug=True)
 
 
@@ -119,4 +115,4 @@ def startwazp():
 # Driver
 ################
 if __name__ == "__main__":
-    startwazp()
+    startbeekeeper()
