+++ docs/source/index.rst
@@ -1,4 +1,4 @@
-.. wazp documentation master file, created by
+.. beekeeper documentation master file, created by
    sphinx-quickstart on Fri Dec  9 14:12:42 2022.
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
