+++ tox.ini
@@ -12,4 +12,4 @@ python =
 extras =
     dev
 commands =
-    pytest -v --color=yes --cov=wazp --cov-report=xml
+    pytest -v --color=yes --cov=beekeeper --cov-report=xml
